Add support for vccd (closed captions) version 2

+++ ValveResourceFormat/ClosedCaptions/ClosedCaption.cs
@@ -6,7 +6,7 @@ namespace ValveResourceFormat.ClosedCaptions
 {
     public class ClosedCaption
     {
-        public uint Hash { get; set; }
+        public ulong Hash { get; set; }
         public int Blocknum { get; set; }
         public ushort Offset { get; set; }
         public ushort Length { get; set; }
